dreaming: Align codebase on dreaming terminology (#338)
Rename "conversation extraction" / "trace review" / "Trace Reviewer" to
"dreaming" / "Dreamer" throughout source and planning docs. The canonical
module is now services/dreaming.py; services/conversation_extraction.py
remains as a backwards-compatibility re-export shim so existing test
patches continue to work.

Guard list preserved: database table names (conversation_extractions,
conversation_extraction_failures), API schema class names, and the
prompts/conversation_extraction.yaml filename are unchanged.
